weights: add ManualWeights
This commit adds methods for managing ManualWeights to the
analysis/weights modules. The manual weights are stored in a map. Any
time an edge evaluator is created, you're now expected to provide the
manual weights as well as the node type weights.

Test plan: Unit tests added, `yarn test` passes.