Add descriptions to the AnalysisAdapter
The descriptions are intended as short (oneline) markdown descriptions.
I'd like to use this as a replacement for the React descriptions in the
explorer adapter. We can generate equivalent descriptions (with
hyperlinks etc) from Markdown, and the benefit of Markdown is that the
data can then be used in other environments, e.g. for export to the
research repo.

This will make it possible to write frontend components that just take
data in, rather than depending on loading complicated objects (e.g.
dynamic adapter sets), which I think will improve testability and
portability of the frontend.

Test plan: See included unit tests, which include snapshots of the
relevant markdown descriptions.