Refactor Graph to support Object Nodes
Presently, the base Graph tracks nodes as `NodeAddressT`s, without the
possibility of storing any metadata with the node.

This commit changes that: now the Graph has a `Node` type, which is an
object. At present, that type only contains an address; it will be
extended in future commits.

See #1136 for some discussion of this approach, and in particular [this
comment] for justification.

[this comment]: https://github.com/sourcecred/sourcecred/issues/1136#issuecomment-498078988

Because this changes the basic APIs in Graph, it churns the entire
codebase. To minimize future churn, I've defined helper functions,
`partsNode` and `node`, in `graphTestUtil`. These helper functions
create simple dummy nodes in situations where we don't much care about
the nodes' fields. This way, as we add new fields we won't need to
update all of the test code again.

Test plan: `yarn test --full` passes. Check the documentation on the
Graph module, and take a look at snapshot changes.